requirements.txt
boto3==1.18.1
botocore==1.21.1
Django==3.2.5
django-anymail==8.4
django-cors-headers==3.7.0
django-rest-framework==0.1.0
django-utils-six==2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
eventlet==0.30.2
greenlet==1.1.0
gunicorn==20.1.0
Pillow==8.3.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
python-dotenv==0.18.0
